requirements.txt
arrow==0.13.1
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
click==7.0
flask==1.0.2
gevent==1.4.0
greenlet==0.4.15 ; platform_python_implementation == 'CPython'
icalendar==4.0.3
idna==2.8
iso8601==0.1.12
itsdangerous==1.1.0
jinja2==2.10
leancloud==2.1.10
lxml==4.3.3
markupsafe==1.1.1
njupt==0.2.3
pillow==5.4.1
python-dateutil==2.8.0
pytz==2018.9
qiniu==7.2.2
requests==2.21.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9
urllib3==1.24.1
werkzeug==0.15.1
